src/requirements.txt
beautifulsoup4==4.11.1
celery==5.2.7
Django==4.1.2
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
latest-user-agents==0.0.3
loguru==0.6.0
pandas==1.5.1
Pillow==9.3.0
psycopg2-binary==2.9.5
redis==4.3.4
requests==2.28.1
sentry-sdk==1.10.1
tqdm==4.64.1